Extract waveform audio extraction module
